Kimlic v1.0.0 AttestationApi.DigitalVerifications

Defines business logic and DB operations with DigitalVerification entity

Functions

create_session(account_address, params)

create_session(binary(), map()) :: {:ok, binary()} | {:error, binary()}

Creates session on Veriff.me and stores verification entity in database

get_by(params)

get_by(map()) ::
  %AttestationApi.DigitalVerifications.DigitalVerification{
    __meta__: term(),
    account_address: term(),
    contract_address: term(),
    device_os: term(),
    device_token: term(),
    document_type: term(),
    documents: term(),
    id: term(),
    inserted_at: term(),
    session_id: term(),
    status: term(),
    updated_at: term(),
    veriffme_code: term(),
    veriffme_reason: term(),
    veriffme_status: term()
  }
  | nil

Finds single DigitalVerification entity by parameters

handle_verification_result(params)

handle_verification_result(map()) :: :ok | {:error, atom()}

Handles webhook from Veriff.me with verification result Updates verification status, sends push notification, removes verification documents and set verification result in Quorum

handle_verification_submission(params)

handle_verification_submission(map()) :: :ok

Handles Veriff.me webhook that notifies about verification submission, updates verification veriffme_code

insert(params)

insert(map()) ::
  {:ok,
   %AttestationApi.DigitalVerifications.DigitalVerification{
     __meta__: term(),
     account_address: term(),
     contract_address: term(),
     device_os: term(),
     device_token: term(),
     document_type: term(),
     documents: term(),
     id: term(),
     inserted_at: term(),
     session_id: term(),
     status: term(),
     updated_at: term(),
     veriffme_code: term(),
     veriffme_reason: term(),
     veriffme_status: term()
   }}
  | {:error, binary()}

Inserts DigitalVerification with parameters

update(entity, params)

update(
  %AttestationApi.DigitalVerifications.DigitalVerification{
    __meta__: term(),
    account_address: term(),
    contract_address: term(),
    device_os: term(),
    device_token: term(),
    document_type: term(),
    documents: term(),
    id: term(),
    inserted_at: term(),
    session_id: term(),
    status: term(),
    updated_at: term(),
    veriffme_code: term(),
    veriffme_reason: term(),
    veriffme_status: term()
  },
  map()
) ::
  {:ok,
   %AttestationApi.DigitalVerifications.DigitalVerification{
     __meta__: term(),
     account_address: term(),
     contract_address: term(),
     device_os: term(),
     device_token: term(),
     document_type: term(),
     documents: term(),
     id: term(),
     inserted_at: term(),
     session_id: term(),
     status: term(),
     updated_at: term(),
     veriffme_code: term(),
     veriffme_reason: term(),
     veriffme_status: term()
   }}
  | {:error, binary()}

Updates DigitalVerification entity with parameters