Kimlic v1.0.0 Core.Clients.Redis

Provides convenient access to Redis using Redix under the hood Serializes and stores data using Erlang term_to_binary

Functions

delete(key)

delete(map()) :: {:ok, non_neg_integer()} | {:error, binary()}

delete(binary()) :: {:ok, non_neg_integer()} | {:error, binary()}

Removes key-value pair by key as:

map that contains redis_key

string

flush()

flush() :: :ok | {:error, binary()}

Removes all data

get(key)

get(binary()) :: {:ok, term()} | {:error, binary()}

Returns value by key

set(key, value, ttl_seconds \\ nil)

set(binary(), term(), pos_integer() | nil) :: :ok | {:error, atom()}

Sets key-value pair

update(entity, params \\ %{}, ttl_seconds \\ nil)

update(struct(), map() | %{}, pos_integer() | nil) ::
  {:ok, term()} | {:error, binary()}

Updates entity as struct with params

upsert(changeset, ttl_seconds \\ nil)

upsert(Ecto.Changeset.t(), pos_integer() | nil) ::
  {:ok, term()} | {:error, binary()}

Updates or inserts changeset which contains redis_key