Kimlic v1.0.0 MobileApi.Router.Helpers

Module with named helpers generated from MobileApi.Router.

Functions

config_path(conn_or_endpoint, action)

config_path(conn_or_endpoint, action, params)

config_url(conn_or_endpoint, action)

config_url(conn_or_endpoint, action, params)

path(data, path)

Generates the path information including any necessary prefix.

push_path(conn_or_endpoint, action)

push_path(conn_or_endpoint, action, params)

push_url(conn_or_endpoint, action)

push_url(conn_or_endpoint, action, params)

quorum_path(conn_or_endpoint, action)

quorum_path(conn_or_endpoint, action, params)

quorum_url(conn_or_endpoint, action)

quorum_url(conn_or_endpoint, action, params)

static_path(conn, path)

Generates path to a static asset given its file path.

static_url(conn, path)

Generates url to a static asset given its file path.

sync_path(conn_or_endpoint, action)

sync_path(conn_or_endpoint, action, params)

sync_url(conn_or_endpoint, action)

sync_url(conn_or_endpoint, action, params)

url(data)

Generates the connection/endpoint base URL without any path information.

verification_path(conn_or_endpoint, action)

verification_path(conn_or_endpoint, action, params)

verification_url(conn_or_endpoint, action)

verification_url(conn_or_endpoint, action, params)